+++ MyPrintiverse/MyPrintiverse.Core/ViewModels/Collections/BaseKeyCollectionWithitemViewModel.cs
@@ -47,14 +47,14 @@ public class BaseKeyCollectionWithitemViewModel<TBaseModel, TEditView, TCollecti
         ItemService = itemService ?? throw new ArgumentNullException(keyItemServiceExceptionMessage);
     }
 
-    protected internal override async void OnAppearing()
+    public override async void OnAppearing()
     {
-        EditDisplayItemCommand = new AsyncCommand(EditDisplayItem, CanExecute);
-        DeleteDisplayItemCommand = new AsyncCommand(DeleteDisplayItem, CanExecute);
+        base.OnAppearing();
+
+        EditDisplayItemCommand = new AsyncCommand(EditDisplayItem, CanExecute, shellExecute: ExecuteBlockade);
+        DeleteDisplayItemCommand = new AsyncCommand(DeleteDisplayItem, CanExecute, shellExecute: ExecuteBlockade);
 
         Item = await ItemService.GetItemAsync(Id);
-
-        base.OnAppearing();
     }
 
     /// <summary>
@@ -69,11 +69,12 @@ public class BaseKeyCollectionWithitemViewModel<TBaseModel, TEditView, TCollecti
     /// <returns></returns>
     protected virtual async Task DeleteDisplayItem()
     {
-        // make sure to delete
+        if (!(await MessageService.ShowSelectAlertAsync("Item Delete", "Do you really want to delete this item?", "Delete")))
+            return;
+
         if (await ItemService.DeleteItemAsync(Item.Id))
             await Shell.Current.GoToAsync("..", true);
 
-        IsBusy = false;
     }
 }
